+++ src/lib/index.ts
@@ -1,6 +1,7 @@
 export * from './command';
 export * from './constant';
 export * from './engine';
+export * from './logger';
 export * from './types';
 export * from './views';